requirements.txt
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.4
django-environ==0.11.2
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
mysqlclient==2.2.0
phonenumbers==8.13.19
python-telegram-bot==20.5
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
